single buffer for creating chunks

+++ src/containers.rs
@@ -22,9 +22,9 @@ impl Inventory {
         }
     }
 
-    pub fn render_item_block(world_data_thread: Arc<Mutex<world::WorldData>>, block_name: String) -> (Vec<[i8; 3]>, Vec<[f32; 2]>, Vec<[i8; 3]>, Vec<[f32; 3]>) {
+    pub fn render_item_block(world_data_thread: Arc<Mutex<world::WorldData>>, block_name: String) -> (Vec<[i64; 3]>, Vec<[f32; 2]>, Vec<[i8; 3]>, Vec<[f32; 3]>) {
         let world_data = world_data_thread.lock().unwrap();
-        let mut vertices: Vec<[i8; 3]> = Vec::new();
+        let mut vertices: Vec<[i64; 3]> = Vec::new();
         let mut normals: Vec<[i8; 3]> = Vec::new();
         let mut colors: Vec<[f32; 3]> = Vec::new();
         let mut uvs: Vec<[f32; 2]> = Vec::new();
